+++ src/TGUI/Widgets/RadioButton.cpp
@@ -126,7 +126,7 @@ namespace tgui
                 m_text.setStyle(getRenderer()->getTextStyle());
 
             m_callback.checked = true;
-            sendSignal("Checked", m_checked);
+            sendSignal("Checked", static_cast<int>(m_checked));
         }
     }
 
@@ -142,7 +142,7 @@ namespace tgui
             m_text.setStyle(getRenderer()->getTextStyle());
 
             m_callback.checked = false;
-            sendSignal("Unchecked", m_checked);
+            sendSignal("Unchecked", static_cast<int>(m_checked));
         }
     }
